requirements.txt
django,
ipython['notebook']
django-extensions
Pillow
PILKit
django-imagekit
django-bootstrap3


